Replace pre generated build files by bootstrap script
The pre generated build files are not working, after a configure,
followed by make, they try a new configure step and request older
versions of automake and libtool, resulting in a build failure. This
patch uses a better solution, which is to remove all generated files and
provide a bootstrap script, which calls autoreconf.